+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `laravel-env-set-command` will be documented in this file.
+
+## [1.0.0] - 2023-11-01
+### Added
+- Laravel 10 support
+- Initial release
+
+This is a copy of Liam Emmett's [package](https://github.com/imliam/laravel-env-set-command)